Extract Direction and Ingredient from Recipe
