+++ src/react/components/Recipe/Components/index.ts
@@ -1,0 +1,4 @@
+import Direction from './Direction';
+import Ingredient from './Ingredient';
+
+export { Direction, Ingredient };
